Use palloc to avoid memleaks

+++ postgis/flatgeobuf.h
@@ -56,7 +56,7 @@ struct flatgeobuf_encode_ctx
 	uint64_t features_count;
 	uint8_t *buf;
 	uint8_t *tmp_buf;
-	uint32_t tmp_buf_size;
+	uint64_t tmp_buf_size;
 	uint64_t offset;
 	uint8_t lwgeom_type;
 	GeometryType_enum_t geometry_type;
